Adicionando exceção  para entradas invalidas no arquivo config.json.

+++ stock-quote-alert/Config/GeneralConfig.cs
@@ -2,8 +2,8 @@
 public class GeneralConfig {
     public MailConfig MailConfig { get; set; }
     public ReportConfig ReportConfig { get; set; }
-    public int DelayToRequest { get; set; }
-    public GeneralConfig(MailConfig mailConfig, int delayToRequest, ReportConfig reportConfig) {
+    public double DelayToRequest { get; set; }
+    public GeneralConfig(MailConfig mailConfig, double delayToRequest, ReportConfig reportConfig) {
         MailConfig = mailConfig;
         ReportConfig = reportConfig;
         DelayToRequest = delayToRequest;
